Fixes #225; see commit body for details (and caveats)
That this issue existed is a side-effect of an earlier chronic issue.  At least
in the pull workflow, no where does the library (on behalf of the client)
attempt to verify the digest returned by the register/server.  By not doing so,
it also has taken an unhealthy dependency on a precariously OPTIONAL response
header, namely, `Docker-Content-Digest`.  All this has lead to issue #225.

In this change, two new functions have been added:

    1) calculateDigestFromResponse
    2) verifyContentDigestForPull

The former calculates the actual digest of the response body.  The latter
performs a 3-way validation between the two potentially-empty digests--namely,
the reference digest (if indeed the reference supplied by the user was a digest
and not a tag), and the response header optionally set by the server (that is,
`Docker-Content-Digest`).

This change--depite possibly being a meager improvement, is still inadequate.
The reason for the inadequacy is that it relies heavily on a response body not
being close, and this is not always the case.  Where it is closed, it again
defaults back to scavenging from any of the two digests available to it, and
neither is guaranteed to even exist:

 1) The header `Docker-Content-Digest`, as already mentioned, is OPTIONAL
 2) The user-supplied `reference` is polymorphic, in that it could just as
    well be a `digest`, as it could a `tag`.

Discussion thread on this has been opened up in Slack:

    https://cloud-native.slack.com/archives/CJ1KHJM5Z/p1657935407555609

Signed-off-by: Nima Talebi <nima@users.noreply.github.com>