Add boolean mode parameter to `generateDescriptor`
The boolean here is `calculateDigest`; when true, the function will insist on
calculating a digest from the response body, and indeed, expect that there is
such a thing as a response body--an open one.

Otherwise it will first try and salvage the digest from the user-supplied
reference (which will fail if it was of the form `:<tag>`, but succeed if it
was instead of the only other alternative form `@<digest>`), and if that fails
it will try and salvage it from the OPTIONAL server response header
`Docker-Content-Digest`.

`verifyContentDigestForPull` was renamed to `verifyContentDigests`.

Signed-off-by: Nima Talebi <nima@users.noreply.github.com>