fix: oras pull error `empty response Docker-Content-Digest` (#237)
That this issue existed is a side-effect of an earlier chronic issue.  At least
in the pull workflow, no where does the library (on behalf of the client)
attempt to verify the digest returned by the register/server.  By not doing so,
it also has taken an unhealthy dependency on a precariously OPTIONAL response
header, namely, `Docker-Content-Digest`.  All this has lead to issue oras-project#225.

Discussion thread on this has been opened up in Slack:

    https://cloud-native.slack.com/archives/CJ1KHJM5Z/p1657935407555609

Signed-off-by: Nima Talebi <github@nima.id.au>