refactor!: change module name to v3 (#1051)
This pull request updates the codebase and documentation to use the new
import path for `oras-go` v3, replacing all references to
`oras.land/oras-go/v2` with `github.com/oras-project/oras-go/v3`. This
resolves #483. The update touches source files, test files, and
documentation.

**Documentation update:**
* The `README.md` has been updated to clarify that the sample code
examples are for `oras-go` v2, improving clarity for users and preparing
for future documentation updates for v3.

---------

Signed-off-by: Shiwei Zhang <shizh@microsoft.com>